+++ src/cairo_run.rs
@@ -1,7 +1,6 @@
 use crate::{
     hint_processor::hint_processor_definition::HintProcessor,
     types::program::Program,
-    utils::RunResources,
     vm::{
         errors::{cairo_run_errors::CairoRunError, vm_exception::VmException},
         runners::cairo_runner::CairoRunner,
@@ -60,7 +59,7 @@ pub fn cairo_run(
     let mut vm = VirtualMachine::new(cairo_run_config.trace_enabled);
     let end = cairo_runner.initialize(&mut vm)?;
     // check step calculation
-    let mut run_resources = Some(RunResources::new(None));
+    let mut run_resources = None;
 
     cairo_runner
         .run_until_pc(end, &mut run_resources, &mut vm, hint_executor)